third_party/blink/web_tests/fast/forms/text/input-text-paste-maxlength-expected.txt

set value attribute that violates maxlength (with pasted value)
PASS domValueOf('f') is '123'
PASS visibleValueOf('f') is '123'
set value property that violates maxlength (with pasted value)
PASS domValueOf('e') is '12345'
PASS visibleValueOf('e') is '12345'
set maxlength attribute that is smaller than pasted value
PASS domValueOf('d') is '12345'
PASS visibleValueOf('d') is '12345'
set maxLength property that is smaller than pasted value
PASS domValueOf('c') is '12345'
PASS visibleValueOf('c') is '12345'
set value attribute that violates maxlength (with pasted value)
PASS domValueOf('j') is '123'
PASS visibleValueOf('j') is '123'
set value property that violates maxlength (with pasted value)
PASS domValueOf('i') is '12' + fancyX + '45'
PASS visibleValueOf('i') is '12' + fancyX + '45'
set maxlength attribute that is smaller than pasted value
PASS domValueOf('h') is '12' + fancyX + '45'
PASS visibleValueOf('h') is '12' + fancyX + '45'
set maxLength property that is smaller than pasted value
PASS domValueOf('g') is '12' + fancyX + '45'
PASS visibleValueOf('g') is '12' + fancyX + '45'
pasting too much text
PASS domValueOf('k-max5') is "12x̲̅"
PASS visibleValueOf('k-max5') is "12x̲̅"
PASS domValueOf('k-max5') is "123𐀀"
PASS visibleValueOf('k-max5') is "123𐀀"
PASS domValueOf('k-max5') is "1234"
PASS visibleValueOf('k-max5') is "1234"
pasting too much text with maxlength=0
PASS domValueOf('l') is ''
PASS visibleValueOf('l') is ''
empty maxlength should be ignored.
PASS domValueOf('m') is '12' + fancyX + '45'
PASS visibleValueOf('m') is '12' + fancyX + '45'
invalid maxlength should be ignored.
PASS domValueOf('n') is '12' + fancyX + '45'
PASS visibleValueOf('n') is '12' + fancyX + '45'
PASS domValueOf('huge') is '12' + fancyX + '45'
PASS visibleValueOf('huge') is '12' + fancyX + '45'
PASS successfullyParsed is true

TEST COMPLETE
